Added `dont_delete_existing_model` Blender option

+++ src/integrations/blender/setup.py
@@ -7,20 +7,26 @@ from src.constants.misc import Executable
 from src.integrations.blender.handlers import initialize_depsgraph_update_handler
 
 
-def setup_blender() -> None:
+def setup_blender(dont_delete_existing_model) -> None:
     if not is_process_running(Executable.BLENDER):
         return
     
-    delete_existing_meshes()
+    delete_existing_meshes(dont_delete_existing_model)
     enable_developer_extras()
     enable_vertex_snapping()
     adjust_3D_view_settings()
     initialize_depsgraph_update_handler()
+    print("Blender setup complete")
 
 
-def delete_existing_meshes() -> None:
+def delete_existing_meshes(dont_delete_existing_model) -> None:
+    if dont_delete_existing_model:
+        print("Existing model not deleted")
+        return
+    
     bpy.ops.object.select_all(action = "SELECT")
     bpy.ops.object.delete()
+    print("Existing model deleted")
 
 
 def enable_developer_extras() -> None:
@@ -30,15 +36,16 @@ def enable_developer_extras() -> None:
     if not view.show_developer_ui:
         view.show_developer_ui = True
         bpy.ops.wm.save_userpref()
-        print("Developer Extras enabled!")
+        print("Developer Extras enabled")
     else:
-        print("Developer Extras already enabled!")
+        print("Developer Extras already enabled")
         
         
 def enable_vertex_snapping() -> None:
     bpy.context.tool_settings.use_snap = True
     bpy.context.tool_settings.snap_elements = {"VERTEX"}
     bpy.context.tool_settings.snap_target = "CLOSEST"  
+    print("Vertex Snapping enabled")
 
 
 def get_3d_space(area: Area) -> Optional[SpaceView3D]:
@@ -50,11 +57,13 @@ def adjust_3D_view_settings() -> None:
         if area.type == "VIEW_3D":
             space = get_3d_space(area)
             if space:
-                space.clip_end = 5000.0  # Clip distance
+                space.clip_end = 5000.0
                
                 shading = space.shading
-                shading.type = "SOLID"  # Set the shading mode to Solid
+                shading.type = "SOLID"
                
-                # Uniform Lighting
                 shading.light = "FLAT"
                 shading.color_type = "TEXTURE"
+                
+                print("3D view settings adjusted")
+                return
